bug squashed: x literals only count if weight above upperKD
When using an x literal to extend the upperKD, I must be careful.
If there are no y variables in the explanation, I need to increase the
upperKD to the minimal weight above the current upperKD. This may not
coincide with the minimal weight, after all the assumptions are
generated by all objective at once. Therefore, there may be one
literal with a small weight present, a weight below the upperKD